Work Completed on 10-2-2018
Work Completed on 10-2-2018 - Completed work on the 'Add Book' form itself, completed the Javascript for handling and submitting the book info, to include the handling of the Custom Fields Extension, and completed some work in the class-wpbooklist-book.php file to actually save a book. Need to modify that file to accept isbn 13s and asin numbers from those new fields in the form. Then need to figure out what needs to be added to Compat.