Work Completed on 10-16-2018
Work Completed on 10-16-2018 - Completed Issue #14 and then worked on Issue #15 - Completed the 'Add a Book' Page/Post functionality. Next need to ensure that the Editing functionality for Pages and Posts is complete - specifically, add a new option dynamically into the 'Create a Page' and 'Create  a Post' drop-downs called 'Already Created' if a Page or Post already exists. If it doesn't already exist, create the page and/or Post upon Editing the book, if the user has choosen 'Yes' in the drop-down.