Work completed on 12-3-2018
Issue #47  - Fixed this - issue was a type (extra space) in the Javascript when trying to retreive the data attribute for the name of the table to pull the books from. This also fixed the Work completed on 12-3-2018 -Searching and Sorting section which was also not working. Also fixed a mix-up in the strings for the Sort by 'Pages (Descending)' and 'Pages (Ascending)' options.